refactor(customer): 고객 목록 상태·액션 로직을 useCustomers hook으로 분리
rel: #88
